cleanup samples: use same structure, only verb expect and more
