+++ samples/js/jasmine/src/test/kotlin/org/atriumlib/samples/jasmine/SampleJsTest.kt
@@ -1,22 +1,20 @@
-import ch.tutteli.atrium.api.fluent.en_GB.messageContains
-import ch.tutteli.atrium.api.fluent.en_GB.toBe
-import ch.tutteli.atrium.api.fluent.en_GB.toThrow
-import ch.tutteli.atrium.api.verbs.assertThat
+package org.atriumlib.samples.jasmine
+
+import ch.tutteli.atrium.api.fluent.en_GB.*
 import ch.tutteli.atrium.api.verbs.expect
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.reporting.RawString
 import ch.tutteli.atrium.translations.DescriptionBasic.IS
 import kotlin.test.Test
 
-class SampleTest {
-
+class SampleJsTest {
     @Test
     fun toBe() {
         expect(1).toBe(1)
     }
 
     @Test
-    fun assertAnExceptionOccurred() {
+    fun expectAnExceptionOccurred() {
         expect {
             throw IllegalArgumentException()
         }.toThrow<IllegalArgumentException>()
@@ -24,7 +22,7 @@ class SampleTest {
 
 
     @Test
-    fun assertAnExceptionWithAMessageOccurred() {
+    fun expectAnExceptionWithAMessageOccurred() {
         expect {
             throw IllegalArgumentException("oho... hello btw")
         }.toThrow<IllegalArgumentException> {
@@ -34,9 +32,8 @@ class SampleTest {
 
     @Test
     fun useOwnFunction() {
-        assertThat(2).isEven()
+        expect(2).isEven()
     }
 }
-
 
 fun Expect<Int>.isEven() = createAndAddAssertion(IS, RawString.create("an even number")) { it % 2 == 0 }
