+++ samples/js/mocha/src/test/kotlin/org/atriumlib/samples/mocha/SampleJsTest.kt
@@ -1,23 +1,20 @@
-package ch.tutteli
+package org.atriumlib.samples.mocha
 
-import ch.tutteli.atrium.api.fluent.en_GB.messageContains
-import ch.tutteli.atrium.api.fluent.en_GB.toBe
-import ch.tutteli.atrium.api.fluent.en_GB.toThrow
-import ch.tutteli.atrium.api.verbs.assertThat
+import ch.tutteli.atrium.api.fluent.en_GB.*
 import ch.tutteli.atrium.api.verbs.expect
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.reporting.RawString
-import ch.tutteli.atrium.translations.DescriptionBasic
-import org.junit.jupiter.api.Test
+import ch.tutteli.atrium.translations.DescriptionBasic.IS
+import kotlin.test.Test
 
-class SampleJvmTest {
+class SampleJsTest {
     @Test
     fun toBe() {
         expect(1).toBe(1)
     }
 
     @Test
-    fun assertAnExceptionOccurred() {
+    fun expectAnExceptionOccurred() {
         expect {
             throw IllegalArgumentException()
         }.toThrow<IllegalArgumentException>()
@@ -25,7 +22,7 @@ class SampleJvmTest {
 
 
     @Test
-    fun assertAnExceptionWithAMessageOccurred() {
+    fun expectAnExceptionWithAMessageOccurred() {
         expect {
             throw IllegalArgumentException("oho... hello btw")
         }.toThrow<IllegalArgumentException> {
@@ -36,10 +33,8 @@ class SampleJvmTest {
     @Test
     fun useOwnFunction() {
         // isEven is defined in the common module
-        assertThat(2).isEven()
+        expect(2).isEven()
     }
 }
 
-fun Expect<Int>.isEven() =
-    createAndAddAssertion(DescriptionBasic.IS, RawString.create("an even number")) { it % 2 == 0 }
-
+fun Expect<Int>.isEven() = createAndAddAssertion(IS, RawString.create("an even number")) { it % 2 == 0 }
