+++ samples/jvm/junit5/build.gradle
@@ -1,28 +1,21 @@
-// Atrium JVM installation: more info here -> https://github.com/robstoll/atrium#jvm
+// Example project to show how to use Atrium in combination with JUnit5
+// For more information on how to setup Atrium for a JVM project -> https://github.com/robstoll/atrium#jvm
 
 import org.gradle.api.tasks.testing.logging.TestExceptionFormat
 
 buildscript {
     ext {
-        kotlin_version= '1.3.71'
-        atrium_version='0.11.0'
-        junit_version= '5.3.1'
-    }
-
-    repositories {
-        jcenter()
-    }
-
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        atrium_version = '0.11.0'
+        junit_version = '5.3.1'
     }
 }
 
 plugins {
-    id("org.jetbrains.kotlin.jvm") version "$kotlin_version"
+    id 'org.jetbrains.kotlin.jvm' version '1.3.72'
 }
 
-apply plugin: 'kotlin'
+group 'org.atriumlib.samples'
+version '0.0.1'
 
 repositories {
     jcenter()
@@ -30,15 +23,15 @@ repositories {
 
 dependencies {
     // junit5 requires jdk8, define dependency to stdlib-jdk at least in test
-    implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version")
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
 
     // setup for Atrium:
     // for infix-api -> change artifact-id to 'atrium-infix-en_GB'
-    testImplementation("ch.tutteli.atrium:atrium-fluent-en_GB:$atrium_version")
+    testImplementation "ch.tutteli.atrium:atrium-fluent-en_GB:$atrium_version"
 
     // setup for junit5
-    testImplementation("org.junit.jupiter:junit-jupiter-api:$junit_version")
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:$junit_version")
+    testImplementation "org.junit.jupiter:junit-jupiter-api:$junit_version"
+    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:$junit_version"
 }
 
 test {
