+++ samples/jvm/spek/build.gradle
@@ -1,28 +1,21 @@
-// Atrium JVM installation: more info here -> https://github.com/robstoll/atrium#jvm
+// Example project to show how to use Atrium in combination with Spek
+// For more information on how to setup Atrium for a JVM project -> https://github.com/robstoll/atrium#jvm
 
 import org.gradle.api.tasks.testing.logging.TestExceptionFormat
 
 buildscript {
     ext {
-        kotlin_version = '1.3.71'
+        atrium_version = '0.11.0'
         spek_version = '2.0.0'
-        atrium_version='0.11.0'
-    }
-
-    repositories {
-        jcenter()
-    }
-
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
 
 plugins {
-    id 'org.jetbrains.kotlin.jvm' version "$kotlin_version"
+    id 'org.jetbrains.kotlin.jvm' version '1.3.72'
 }
 
-apply plugin: 'kotlin'
+group 'org.atriumlib.samples'
+version '0.0.1'
 
 repositories {
     jcenter()
@@ -30,7 +23,7 @@ repositories {
 
 dependencies {
     //spek requires jdk8, define dependency to stdlib-jdk at least in test
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
 
     // setup for Atrium:
     // for infix-api -> change artifact-id to 'atrium-infix-en_GB'
@@ -39,9 +32,6 @@ dependencies {
     // setup for 'spek', for some kotlin-version
     testImplementation "org.spekframework.spek2:spek-dsl-jvm:$spek_version"
     testRuntimeOnly "org.spekframework.spek2:spek-runner-junit5:$spek_version"
-
-    // spek requires kotlin-reflect, can be omitted if already in the classpath
-    testRuntimeOnly "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
 }
 
 test {
