+++ samples/jvm/spek/src/main/kotlin/org/atriumlib/samples/spek/Calculator.kt
@@ -1,4 +1,4 @@
-package org.spekframework.spek2
+package org.atriumlib.samples.spek
 
 class Calculator {
     val type = "fancy"
