+++ samples/multiplatform/build.gradle
@@ -1,42 +1,70 @@
+// Example project to show how to use Atrium in a multiplatform project
+// For more information on how to setup Atrium for a multiplatform project -> https://github.com/robstoll/atrium#common
+
+import org.gradle.api.tasks.testing.logging.TestExceptionFormat
+
+buildscript {
+    ext {
+        // for infix-api -> change to 'atrium-infix-en_GB'
+        atrium_api = 'atrium-fluent-en_GB'
+        atrium_version = '0.11.0'
+        junit_version = '5.3.1'
+    }
+}
+
 plugins {
-    id 'org.jetbrains.kotlin.multiplatform' version '1.3.71'
+    id 'org.jetbrains.kotlin.multiplatform' version '1.3.72'
 }
-repositories {
-    mavenCentral()
-    jcenter()
-}
-group 'com.atrium'
+
+group 'org.atriumlib.samples'
 version '0.0.1'
 
-apply plugin: 'maven-publish'
+repositories {
+    jcenter()
+}
 
 kotlin {
-    def atrium_version = '0.11.0'
     jvm()
-    js {
-        nodejs {}
-    }
+    js().nodejs{}
     sourceSets {
         commonTest {
             dependencies {
+                // setup for Atrium:
+                implementation "ch.tutteli.atrium:$atrium_api-common:$atrium_version"
+
+                // setup for common tests:
                 implementation kotlin('test-common')
                 implementation kotlin('test-annotations-common')
-                implementation "ch.tutteli.atrium:atrium-fluent-en_GB-common:$atrium_version"
             }
         }
         jvmTest {
             dependencies {
+                // setup for Atrium:
+                implementation "ch.tutteli.atrium:$atrium_api:$atrium_version"
+
+                // setup for Junit5:
                 implementation kotlin('test')
                 implementation kotlin('test-junit')
-                implementation "org.junit.jupiter:junit-jupiter-api:5.3.1"
-                implementation "ch.tutteli.atrium:atrium-fluent-en_GB:$atrium_version"
+                implementation "org.junit.jupiter:junit-jupiter-api:$junit_version"
             }
         }
         jsTest {
             dependencies {
+                // setup for Atrium:
+                implementation "ch.tutteli.atrium:$atrium_api-js:$atrium_version"
+
+                // setup for mocha:
                 implementation kotlin('test-js')
-                implementation "ch.tutteli.atrium:atrium-fluent-en_GB-js:$atrium_version"
             }
         }
+    }
+}
+
+jvmTest {
+    testLogging {
+        exceptionFormat TestExceptionFormat.FULL
+        showExceptions true
+        showCauses true
+        showStackTraces true
     }
 }
